+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [2-beta21] - 2024-10-27
+- updated deps: kindly-advice
+- added for expansion (`emmy.viewer/expand`) for `:kind/emmy-viewers`
+
 ## [2-beta20] - 2024-10-22
 - updated deps: kindly, kindly-advice
-- added support for `:kind/emmy-viewers`
+- added partial support for `:kind/emmy-viewers`
 
 ## [2-beta19] - 2024-10-22
 - a minor cleanup of the data structure being read from code